# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TaskFlow API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TaskPriority = (typeof TaskPriority)[keyof typeof TaskPriority];

export const TaskPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export type TaskStatus = (typeof TaskStatus)[keyof typeof TaskStatus];

export const TaskStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
} as const;

export interface Task {
  id: number;
  userId: string;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  dueDate?: string | null;
  priority: TaskPriority;
  status: TaskStatus;
  /** @nullable */
  category?: string | null;
  tags: string[];
  createdAt: string;
  updatedAt: string;
}

export type CreateTaskBodyPriority =
  (typeof CreateTaskBodyPriority)[keyof typeof CreateTaskBodyPriority];

export const CreateTaskBodyPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export type CreateTaskBodyStatus =
  (typeof CreateTaskBodyStatus)[keyof typeof CreateTaskBodyStatus];

export const CreateTaskBodyStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
} as const;

export interface CreateTaskBody {
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  dueDate?: string | null;
  priority: CreateTaskBodyPriority;
  status: CreateTaskBodyStatus;
  /** @nullable */
  category?: string | null;
  tags?: string[];
}

export type UpdateTaskBodyPriority =
  (typeof UpdateTaskBodyPriority)[keyof typeof UpdateTaskBodyPriority];

export const UpdateTaskBodyPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export type UpdateTaskBodyStatus =
  (typeof UpdateTaskBodyStatus)[keyof typeof UpdateTaskBodyStatus];

export const UpdateTaskBodyStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
} as const;

export interface UpdateTaskBody {
  title?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  dueDate?: string | null;
  priority?: UpdateTaskBodyPriority;
  status?: UpdateTaskBodyStatus;
  /** @nullable */
  category?: string | null;
  tags?: string[];
}

export interface StatsSummary {
  total: number;
  pending: number;
  inProgress: number;
  completed: number;
  overdue: number;
  completionRate: number;
}

export interface PriorityCount {
  priority: string;
  count: number;
}

export interface StatusCount {
  status: string;
  count: number;
}

export type GetTasksParams = {
  status?: GetTasksStatus;
  priority?: GetTasksPriority;
  search?: string;
  category?: string;
  dueBefore?: string;
  dueAfter?: string;
};

export type GetTasksStatus =
  (typeof GetTasksStatus)[keyof typeof GetTasksStatus];

export const GetTasksStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
} as const;

export type GetTasksPriority =
  (typeof GetTasksPriority)[keyof typeof GetTasksPriority];

export const GetTasksPriority = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;