# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TaskFlow API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: tasks
    description: Task management operations
  - name: stats
    description: Dashboard statistics
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /tasks:
    get:
      operationId: getTasks
      tags: [tasks]
      summary: List tasks for the current user
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [pending, in_progress, completed]
        - name: priority
          in: query
          required: false
          schema:
            type: string
            enum: [low, medium, high]
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: dueBefore
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: dueAfter
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
        "401":
          description: Unauthorized
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Invalid input
        "401":
          description: Unauthorized

  /tasks/{id}:
    get:
      operationId: getTask
      tags: [tasks]
      summary: Get a single task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
        "401":
          description: Unauthorized
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskBody"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "400":
          description: Invalid input
        "404":
          description: Not found
        "401":
          description: Unauthorized
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
        "401":
          description: Unauthorized

  /tasks/{id}/complete:
    patch:
      operationId: toggleTaskComplete
      tags: [tasks]
      summary: Toggle task completion status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "404":
          description: Not found
        "401":
          description: Unauthorized

  /stats/summary:
    get:
      operationId: getStatsSummary
      tags: [stats]
      summary: Get task summary statistics for the current user
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsSummary"
        "401":
          description: Unauthorized

  /stats/by-priority:
    get:
      operationId: getStatsByPriority
      tags: [stats]
      summary: Get task counts grouped by priority
      responses:
        "200":
          description: Priority breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PriorityCount"
        "401":
          description: Unauthorized

  /stats/by-status:
    get:
      operationId: getStatsByStatus
      tags: [stats]
      summary: Get task counts grouped by status
      responses:
        "200":
          description: Status breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StatusCount"
        "401":
          description: Unauthorized

  /tasks/upcoming:
    get:
      operationId: getUpcomingTasks
      tags: [tasks]
      summary: Get tasks due in the next 7 days
      responses:
        "200":
          description: Upcoming tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
        "401":
          description: Unauthorized

  /tasks/overdue:
    get:
      operationId: getOverdueTasks
      tags: [tasks]
      summary: Get overdue tasks
      responses:
        "200":
          description: Overdue tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
        "401":
          description: Unauthorized

  /categories:
    get:
      operationId: getCategories
      tags: [tasks]
      summary: Get all unique categories for the current user
      responses:
        "200":
          description: Categories list
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        "401":
          description: Unauthorized

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Task:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
          format: date-time
        priority:
          type: string
          enum: [low, medium, high]
        status:
          type: string
          enum: [pending, in_progress, completed]
        category:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - priority
        - status
        - tags
        - createdAt
        - updatedAt

    CreateTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
          format: date-time
        priority:
          type: string
          enum: [low, medium, high]
        status:
          type: string
          enum: [pending, in_progress, completed]
        category:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
      required:
        - title
        - priority
        - status

    UpdateTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
          format: date-time
        priority:
          type: string
          enum: [low, medium, high]
        status:
          type: string
          enum: [pending, in_progress, completed]
        category:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string

    StatsSummary:
      type: object
      properties:
        total:
          type: integer
        pending:
          type: integer
        inProgress:
          type: integer
        completed:
          type: integer
        overdue:
          type: integer
        completionRate:
          type: number
      required:
        - total
        - pending
        - inProgress
        - completed
        - overdue
        - completionRate

    PriorityCount:
      type: object
      properties:
        priority:
          type: string
        count:
          type: integer
      required:
        - priority
        - count

    StatusCount:
      type: object
      properties:
        status:
          type: string
        count:
          type: integer
      required:
        - status
        - count