# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TaskFlow API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List tasks for the current user
 */
export const GetTasksQueryParams = zod.object({
  status: zod.enum(["pending", "in_progress", "completed"]).optional(),
  priority: zod.enum(["low", "medium", "high"]).optional(),
  search: zod.coerce.string().optional(),
  category: zod.coerce.string().optional(),
  dueBefore: zod.date().optional(),
  dueAfter: zod.date().optional(),
});

export const GetTasksResponseItem = zod.object({
  id: zod.number(),
  userId: zod.string(),
  title: zod.string(),
  description: zod.string().nullish(),
  dueDate: zod.coerce.date().nullish(),
  priority: zod.enum(["low", "medium", "high"]),
  status: zod.enum(["pending", "in_progress", "completed"]),
  category: zod.string().nullish(),
  tags: zod.array(zod.string()),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const GetTasksResponse = zod.array(GetTasksResponseItem);

/**
 * @summary Create a new task
 */
export const CreateTaskBody = zod.object({
  title: zod.string(),
  description: zod.string().nullish(),
  dueDate: zod.coerce.date().nullish(),
  priority: zod.enum(["low", "medium", "high"]),
  status: zod.enum(["pending", "in_progress", "completed"]),
  category: zod.string().nullish(),
  tags: zod.array(zod.string()).optional(),
});

/**
 * @summary Get a single task
 */
export const GetTaskParams = zod.object({
  id: zod.coerce.number(),
});

export const GetTaskResponse = zod.object({
  id: zod.number(),
  userId: zod.string(),
  title: zod.string(),
  description: zod.string().nullish(),
  dueDate: zod.coerce.date().nullish(),
  priority: zod.enum(["low", "medium", "high"]),
  status: zod.enum(["pending", "in_progress", "completed"]),
  category: zod.string().nullish(),
  tags: zod.array(zod.string()),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Update a task
 */
export const UpdateTaskParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateTaskBody = zod.object({
  title: zod.string().optional(),
  description: zod.string().nullish(),
  dueDate: zod.coerce.date().nullish(),
  priority: zod.enum(["low", "medium", "high"]).optional(),
  status: zod.enum(["pending", "in_progress", "completed"]).optional(),
  category: zod.string().nullish(),
  tags: zod.array(zod.string()).optional(),
});

export const UpdateTaskResponse = zod.object({
  id: zod.number(),
  userId: zod.string(),
  title: zod.string(),
  description: zod.string().nullish(),
  dueDate: zod.coerce.date().nullish(),
  priority: zod.enum(["low", "medium", "high"]),
  status: zod.enum(["pending", "in_progress", "completed"]),
  category: zod.string().nullish(),
  tags: zod.array(zod.string()),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Delete a task
 */
export const DeleteTaskParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Toggle task completion status
 */
export const ToggleTaskCompleteParams = zod.object({
  id: zod.coerce.number(),
});

export const ToggleTaskCompleteResponse = zod.object({
  id: zod.number(),
  userId: zod.string(),
  title: zod.string(),
  description: zod.string().nullish(),
  dueDate: zod.coerce.date().nullish(),
  priority: zod.enum(["low", "medium", "high"]),
  status: zod.enum(["pending", "in_progress", "completed"]),
  category: zod.string().nullish(),
  tags: zod.array(zod.string()),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Get task summary statistics for the current user
 */
export const GetStatsSummaryResponse = zod.object({
  total: zod.number(),
  pending: zod.number(),
  inProgress: zod.number(),
  completed: zod.number(),
  overdue: zod.number(),
  completionRate: zod.number(),
});

/**
 * @summary Get task counts grouped by priority
 */
export const GetStatsByPriorityResponseItem = zod.object({
  priority: zod.string(),
  count: zod.number(),
});
export const GetStatsByPriorityResponse = zod.array(
  GetStatsByPriorityResponseItem,
);

/**
 * @summary Get task counts grouped by status
 */
export const GetStatsByStatusResponseItem = zod.object({
  status: zod.string(),
  count: zod.number(),
});
export const GetStatsByStatusResponse = zod.array(GetStatsByStatusResponseItem);

/**
 * @summary Get tasks due in the next 7 days
 */
export const GetUpcomingTasksResponseItem = zod.object({
  id: zod.number(),
  userId: zod.string(),
  title: zod.string(),
  description: zod.string().nullish(),
  dueDate: zod.coerce.date().nullish(),
  priority: zod.enum(["low", "medium", "high"]),
  status: zod.enum(["pending", "in_progress", "completed"]),
  category: zod.string().nullish(),
  tags: zod.array(zod.string()),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const GetUpcomingTasksResponse = zod.array(GetUpcomingTasksResponseItem);

/**
 * @summary Get overdue tasks
 */
export const GetOverdueTasksResponseItem = zod.object({
  id: zod.number(),
  userId: zod.string(),
  title: zod.string(),
  description: zod.string().nullish(),
  dueDate: zod.coerce.date().nullish(),
  priority: zod.enum(["low", "medium", "high"]),
  status: zod.enum(["pending", "in_progress", "completed"]),
  category: zod.string().nullish(),
  tags: zod.array(zod.string()),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const GetOverdueTasksResponse = zod.array(GetOverdueTasksResponseItem);

/**
 * @summary Get all unique categories for the current user
 */
export const GetCategoriesResponseItem = zod.string();
export const GetCategoriesResponse = zod.array(GetCategoriesResponseItem);