# lib/api-zod/src/generated/types/createTaskBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TaskFlow API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateTaskBodyPriority } from "./createTaskBodyPriority";
import type { CreateTaskBodyStatus } from "./createTaskBodyStatus";

export interface CreateTaskBody {
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  dueDate?: Date | null;
  priority: CreateTaskBodyPriority;
  status: CreateTaskBodyStatus;
  /** @nullable */
  category?: string | null;
  tags?: string[];
}