# lib/api-zod/src/generated/types/priorityCount.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TaskFlow API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PriorityCount {
  priority: string;
  count: number;
}