# lib/api-zod/src/generated/types/task.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TaskFlow API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TaskPriority } from "./taskPriority";
import type { TaskStatus } from "./taskStatus";

export interface Task {
  id: number;
  userId: string;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  dueDate?: Date | null;
  priority: TaskPriority;
  status: TaskStatus;
  /** @nullable */
  category?: string | null;
  tags: string[];
  createdAt: Date;
  updatedAt: Date;
}